python_scripts/requirements.txt
requests
paho-mqtt
krakenex
PyYAML
